requirements.txt
torch
transformers
grpcio
grpcio-tools